Implement merge of config and credentials file
Loading the profile data from the config and credentials files
will new merge the data of the two files. The data in the second
file is preferred. This means in the current load order that the
credentials file will overwrite configuration in the config
file.

Test data and test was added for this case.